Move condition expressions into own package
