+++ h2/src/main/org/h2/expression/Parameter.java
@@ -7,6 +7,7 @@ package org.h2.expression;
 
 import org.h2.api.ErrorCode;
 import org.h2.engine.Session;
+import org.h2.expression.condition.Comparison;
 import org.h2.message.DbException;
 import org.h2.table.Column;
 import org.h2.table.ColumnResolver;
