+++ h2/src/main/org/h2/expression/ValueExpression.java
@@ -6,6 +6,7 @@
 package org.h2.expression;
 
 import org.h2.engine.Session;
+import org.h2.expression.condition.Comparison;
 import org.h2.index.IndexCondition;
 import org.h2.message.DbException;
 import org.h2.table.ColumnResolver;
