+++ h2/src/main/org/h2/expression/condition/Condition.java
@@ -3,8 +3,9 @@
  * and the EPL 1.0 (http://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
-package org.h2.expression;
+package org.h2.expression.condition;
 
+import org.h2.expression.Expression;
 import org.h2.value.Value;
 import org.h2.value.ValueBoolean;
