+++ h2/src/main/org/h2/expression/condition/ConditionAndOr.java
@@ -3,10 +3,13 @@
  * and the EPL 1.0 (http://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
-package org.h2.expression;
+package org.h2.expression.condition;
 
 import org.h2.engine.Session;
 import org.h2.engine.SysProperties;
+import org.h2.expression.Expression;
+import org.h2.expression.ExpressionVisitor;
+import org.h2.expression.ValueExpression;
 import org.h2.message.DbException;
 import org.h2.table.ColumnResolver;
 import org.h2.table.TableFilter;
