+++ h2/src/main/org/h2/index/IndexCondition.java
@@ -11,10 +11,10 @@ import java.util.HashSet;
 import java.util.List;
 import org.h2.command.dml.Query;
 import org.h2.engine.Session;
-import org.h2.expression.Comparison;
 import org.h2.expression.Expression;
 import org.h2.expression.ExpressionColumn;
 import org.h2.expression.ExpressionVisitor;
+import org.h2.expression.condition.Comparison;
 import org.h2.message.DbException;
 import org.h2.result.ResultInterface;
 import org.h2.table.Column;
