+++ h2/src/main/org/h2/index/ViewIndex.java
@@ -16,8 +16,8 @@ import org.h2.command.dml.Query;
 import org.h2.command.dml.SelectUnion;
 import org.h2.engine.Constants;
 import org.h2.engine.Session;
-import org.h2.expression.Comparison;
 import org.h2.expression.Parameter;
+import org.h2.expression.condition.Comparison;
 import org.h2.message.DbException;
 import org.h2.result.LocalResult;
 import org.h2.result.ResultInterface;
