+++ h2/src/main/org/h2/table/Column.java
@@ -15,11 +15,11 @@ import org.h2.engine.Constants;
 import org.h2.engine.Domain;
 import org.h2.engine.Mode;
 import org.h2.engine.Session;
-import org.h2.expression.ConditionAndOr;
 import org.h2.expression.Expression;
 import org.h2.expression.ExpressionVisitor;
 import org.h2.expression.SequenceValue;
 import org.h2.expression.ValueExpression;
+import org.h2.expression.condition.ConditionAndOr;
 import org.h2.message.DbException;
 import org.h2.result.Row;
 import org.h2.schema.Schema;
