+++ h2/src/test/org/h2/test/unit/TestPattern.java
@@ -6,7 +6,7 @@
 package org.h2.test.unit;
 
 import java.text.Collator;
-import org.h2.expression.CompareLike;
+import org.h2.expression.condition.CompareLike;
 import org.h2.test.TestBase;
 import org.h2.value.CompareMode;
